Ákos féle változat módosítása

+++ final-exam-application-2/final-exam-application-2-service/src/main/java/hu/icell/javaeetraining/finalexam/application2/service/TodoTaskProxy.java
@@ -4,7 +4,9 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
 
-import javax.inject.Inject;
+import javax.ejb.EJB;
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Named;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -20,20 +22,21 @@ import javax.ws.rs.core.UriInfo;
 import hu.icell.javaeetraining.finalexam.application2.dto.TaskDTO;
 import hu.icell.javaeetraining.finalexam.application2.dto.TodoDTO;
 
-@Path("/service")
+@Path("/todo")
+@Named
+@RequestScoped
 public class TodoTaskProxy {
 
 	@Context
 	private UriInfo uriInfo;
 	
-	@Inject
-	private TodoTaskService service;
+	@EJB
+	private TodoService service;
 	
 	@PUT
-	@Path("/todo")
 	@Consumes(MediaType.APPLICATION_JSON)
 	public Response createTodo(TodoDTO todo) throws URISyntaxException {
-		
+
 		Integer todoID = service.createTodo(todo);
 
 		URI location = new URI(uriInfo.getRequestUri() + "/" + todoID);
@@ -41,7 +44,7 @@ public class TodoTaskProxy {
 	}
 
 	@DELETE
-	@Path("/todo/{todoID}")
+	@Path("/{todoID}")
 	public Response deleteTodo(@PathParam("todoID") Integer todoID) {
 		
 		service.deleteTodo(todoID);
@@ -50,26 +53,34 @@ public class TodoTaskProxy {
 	}
 	
 	@PUT
-	@Path("/task")
+	@Path("/{todoID}/task")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response createTask(TaskDTO task) {
+	public Response createTask(@PathParam("todoID") Integer todoID, TaskDTO task) {
 		
+		task.setTodoId(todoID);
 		service.addTask(task);
 		
 		return Response.status(Response.Status.CREATED).entity(null).build();
 	}
 
 	@DELETE
-	@Path("/task/{taskID}")
-	public Response deleteTask(@PathParam("taskID") Integer taskID) {
+	@Path("/{todoID}/task/{taskID}")
+	public Response deleteTask(@PathParam("todoID") Integer todoID, @PathParam("taskID") Integer taskID) {
 		
-		service.removeTask(taskID);
-		
-		return Response.status(Response.Status.OK).entity(null).build();
+		List<TaskDTO> tasks = service.getAllTaskForTodo(todoID);
+		for (TaskDTO task : tasks) {
+			if (task.getId() == taskID) {
+				service.removeTask(taskID);
+				return Response.status(Response.Status.OK).entity(null).build();
+			}
+		}
+
+		return Response.status(Response.Status.NOT_FOUND).
+			header("message", "No such task for this todo. todoId: "+todoID+", taskId: "+taskID)
+			.build();
 	}
 	
 	@GET
-	@Path("/todo")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response listTodos() {
 		
@@ -79,7 +90,7 @@ public class TodoTaskProxy {
 	}
 	
 	@GET
-	@Path("/todo/{todoID}")
+	@Path("/{todoID}/task")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response listTodoTasks(@PathParam("todoID") Integer todoID) {
